--- urdfenvs/urdf_common/__init__.py
@@ -1,6 +1,0 @@
-from gymnasium import register
-from urdfenvs.urdf_common.urdf_env import UrdfEnv
-register(
-    id='urdf-env-v0',
-    entry_point='urdfenvs.urdf_common:UrdfEnv'
-)
