+++ urdfenvs/urdf_common/plane.py
@@ -1,10 +1,10 @@
-import pybullet as p
 import os
+import pybullet
 
 
 class Plane:
     def __init__(self):
         f_name = os.path.join(os.path.dirname(__file__), 'plane.urdf')
-        p.loadURDF(f_name)
+        pybullet.loadURDF(f_name)
